+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Render asset field with multiple values as grid
 - Add support for ${ENV_VAR} in config values
 - Allow custom database settings on space creation
+- Spaces groups
 
 ## 2.7.2 (2023-12-30)